services/wifi-pose/requirements.txt
paho-mqtt>=2.0
numpy>=1.24
opencv-python-headless>=4.8
pyyaml>=6.0
